moved rest Config builder to common package
On-behalf-of: @SAP a.shcherbatiuk@sap.com
